requirements.txt
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.16
django-allauth==65.3.1
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
